Fixed nasty bug in get_40 accessor
Inadvertent sign-extension meant that if the least significant byte of
the value was > 0x7F, the upper 32 bits of the result would fill with
0xFF bytes.

This was screwing up the values computed for the document count and the
by-sequence count.

Fixed the bug and added unit tests for get_40 and get_48 (and the
corresponding use of set_bits.)

Reviewed-on: http://review.couchbase.org/14979
Tested-by: Damien Katz <damien@couchbase.com>

+++ src/bitfield.h
@@ -16,7 +16,7 @@ static inline uint64_t get_48(const char *buf)
 static inline uint64_t get_40(const char *buf)
 {
     const uint32_t* longs = (const uint32_t*)buf;
-    return ((uint64_t)ntohl(longs[0]) << 8) | buf[4];
+    return ((uint64_t)ntohl(longs[0]) << 8) | (uint8_t)buf[4];
 }
 
 /** Read a 32-bit big-endian integer from the address pointed to by buf. */
